+++ tests/actions/test/test.lua
@@ -1,4 +1,4 @@
 function main(t)
-    os.exec("xmake test")
+    os.exec("xmake test -vD")
 end